Conflicts:
	src/CMakeLists.txt

+++ src/p_map.cpp
@@ -329,8 +329,7 @@ bool P_TeleportMove (AActor *thing, fixed_t x, fixed_t y, fixed_t z, bool telefr
 					
 	spechit.Clear ();
 
-	bool StompAlwaysFrags = (thing->flags2 & MF2_TELESTOMP) || 
-							(level.flags & LEVEL_MONSTERSTELEFRAG) || telefrag;
+	bool StompAlwaysFrags = ((thing->flags2 & MF2_TELESTOMP) || (level.flags & LEVEL_MONSTERSTELEFRAG) || telefrag) && !(thing->flags7 & MF7_NOTELESTOMP);
 
 	FBoundingBox box(x, y, thing->radius);
 	FBlockLinesIterator it(box);
@@ -378,7 +377,8 @@ bool P_TeleportMove (AActor *thing, fixed_t x, fixed_t y, fixed_t z, bool telefr
 
 		// monsters don't stomp things except on boss level
 		// [RH] Some Heretic/Hexen monsters can telestomp
-		if (StompAlwaysFrags && !(th->flags6 & MF6_NOTELEFRAG))
+		// ... and some items can never be telefragged while others will be telefragged by everything that teleports upon them.
+		if ((StompAlwaysFrags && !(th->flags6 & MF6_NOTELEFRAG)) || (th->flags7 & MF7_ALWAYSTELEFRAG))
 		{
 			P_DamageMobj (th, thing, thing, TELEFRAG_DAMAGE, NAME_Telefrag, DMG_THRUSTLESS);
 			continue;
@@ -914,34 +914,34 @@ bool PIT_CheckThing (AActor *thing, FCheckPosition &tm)
 		}
 	}
 
-	// touchy object is alive, toucher is solid
-	if (thing->flags6 & MF6_TOUCHY && tm.thing->flags & MF_SOLID && thing->health > 0 &&
-	// Thing is an armed mine or a sentient thing
-	(thing->flags6 & MF6_ARMED || thing->IsSentient()) &&
-	// either different classes or players
-		(thing->player || thing->GetClass() != tm.thing->GetClass()) &&
-	// or different species if DONTHARMSPECIES
-		(!(thing->flags6 & MF6_DONTHARMSPECIES) || thing->GetSpecies() != tm.thing->GetSpecies()) &&
-	// touches vertically
-		thing->z + thing->height >= tm.thing->z && tm.thing->z + tm.thing->height >= thing->z &&
-	// prevents lost souls from exploding when fired by pain elementals
-		(thing->master != tm.thing && tm.thing->master != thing))
-	// Difference with MBF: MBF hardcodes the LS/PE check and lets actors of the same species
-	// but different classes trigger the touchiness, but that seems less straightforwards.
+	if (tm.thing->player == NULL || !(tm.thing->player->cheats & CF_PREDICTING))
 	{
-		thing->flags6 &= ~MF6_ARMED; // Disarm
-		P_DamageMobj (thing, NULL, NULL, thing->health, NAME_None, DMG_FORCED);  // kill object
-		return true;
-	}
+		// touchy object is alive, toucher is solid
+		if (thing->flags6 & MF6_TOUCHY && tm.thing->flags & MF_SOLID && thing->health > 0 &&
+			// Thing is an armed mine or a sentient thing
+			(thing->flags6 & MF6_ARMED || thing->IsSentient()) &&
+			// either different classes or players
+			(thing->player || thing->GetClass() != tm.thing->GetClass()) &&
+			// or different species if DONTHARMSPECIES
+			(!(thing->flags6 & MF6_DONTHARMSPECIES) || thing->GetSpecies() != tm.thing->GetSpecies()) &&
+			// touches vertically
+			thing->z + thing->height >= tm.thing->z && tm.thing->z + tm.thing->height >= thing->z &&
+			// prevents lost souls from exploding when fired by pain elementals
+			(thing->master != tm.thing && tm.thing->master != thing))
+			// Difference with MBF: MBF hardcodes the LS/PE check and lets actors of the same species
+			// but different classes trigger the touchiness, but that seems less straightforwards.
+		{
+			thing->flags6 &= ~MF6_ARMED; // Disarm
+			P_DamageMobj(thing, NULL, NULL, thing->health, NAME_None, DMG_FORCED);  // kill object
+			return true;
+		}
 
-	// Check for MF6_BUMPSPECIAL
-	// By default, only players can activate things by bumping into them
-	if ((thing->flags6 & MF6_BUMPSPECIAL) && ((tm.thing->player != NULL)
-		|| ((thing->activationtype & THINGSPEC_MonsterTrigger) && (tm.thing->flags3 & MF3_ISMONSTER))
-		|| ((thing->activationtype & THINGSPEC_MissileTrigger) && (tm.thing->flags & MF_MISSILE))
-		) && (level.maptime > thing->lastbump)) // Leave the bumper enough time to go away
-	{
-		if (tm.thing->player == NULL || !(tm.thing->player->cheats & CF_PREDICTING))
+		// Check for MF6_BUMPSPECIAL
+		// By default, only players can activate things by bumping into them
+		if ((thing->flags6 & MF6_BUMPSPECIAL) && ((tm.thing->player != NULL)
+			|| ((thing->activationtype & THINGSPEC_MonsterTrigger) && (tm.thing->flags3 & MF3_ISMONSTER))
+			|| ((thing->activationtype & THINGSPEC_MissileTrigger) && (tm.thing->flags & MF_MISSILE))
+			) && (level.maptime > thing->lastbump)) // Leave the bumper enough time to go away
 		{
 			if (P_ActivateThingSpecial(thing, tm.thing))
 				thing->lastbump = level.maptime + TICRATE;
@@ -955,6 +955,17 @@ bool PIT_CheckThing (AActor *thing, FCheckPosition &tm)
 		tm.thing->BlockingMobj = NULL;
 		return res;
 	}
+
+	// [ED850] Player Prediction ends here. There is nothing else they could/should do.
+	if (tm.thing->player != NULL && (tm.thing->player->cheats & CF_PREDICTING))
+	{
+		solid = (thing->flags & MF_SOLID) &&
+			!(thing->flags & MF_NOCLIP) &&
+			((tm.thing->flags & MF_SOLID) || (tm.thing->flags6 & MF6_BLOCKEDBYSOLIDACTORS));
+
+		return !solid || unblocking;
+	}
+
 	// Check for blasted thing running into another
 	if ((tm.thing->flags2 & MF2_BLASTED) && (thing->flags & MF_SHOOTABLE))
 	{
@@ -1203,8 +1214,7 @@ bool PIT_CheckThing (AActor *thing, FCheckPosition &tm)
 		}
 		return false;		// don't traverse any more
 	}
-	if (thing->flags2 & MF2_PUSHABLE && !(tm.thing->flags2 & MF2_CANNOTPUSH) &&
-		(tm.thing->player == NULL || !(tm.thing->player->cheats & CF_PREDICTING)))
+	if (thing->flags2 & MF2_PUSHABLE && !(tm.thing->flags2 & MF2_CANNOTPUSH))
 	{ // Push thing
 		if (thing->lastpush != tm.PushTime)
 		{
@@ -1640,7 +1650,7 @@ static void CheckForPushSpecial (line_t *line, int side, AActor *mobj, bool wind
 {
 	if (line->special && !(mobj->flags6 & MF6_NOTRIGGER))
 	{
-		if (windowcheck && line->backsector != NULL)
+		if (windowcheck && !(ib_compatflags & BCOMPATF_NOWINDOWCHECK) && line->backsector != NULL)
 		{ // Make sure this line actually blocks us and is not a window
 		  // or similar construct we are standing inside of.
 			fixed_t fzt = line->frontsector->ceilingplane.ZatPoint(mobj->x, mobj->y);
@@ -4879,7 +4889,8 @@ void P_DoCrunch (AActor *thing, FChangePosition *cpos)
 				const PClass *bloodcls = thing->GetBloodType();
 				
 				P_TraceBleed (newdam > 0 ? newdam : cpos->crushchange, thing);
-				if (cl_bloodtype <= 1 && bloodcls != NULL)
+
+				if (bloodcls != NULL)
 				{
 					AActor *mo;
 
@@ -4892,14 +4903,15 @@ void P_DoCrunch (AActor *thing, FChangePosition *cpos)
 					{
 						mo->Translation = TRANSLATION(TRANSLATION_Blood, bloodcolor.a);
 					}
+
+					if (!(cl_bloodtype <= 1)) mo->renderflags |= RF_INVISIBLE;
 				}
+
+				angle_t an;
+				an = (M_Random () - 128) << 24;
 				if (cl_bloodtype >= 1)
 				{
-					angle_t an;
-
-					an = (M_Random () - 128) << 24;
-					P_DrawSplash2 (32, thing->x, thing->y,
-								   thing->z + thing->height/2, an, 2, bloodcolor);
+					P_DrawSplash2(32, thing->x, thing->y, thing->z + thing->height / 2, an, 2, bloodcolor);
 				}
 			}
 			if (thing->CrushPainSound != 0 && !S_GetSoundPlayingInfo(thing, thing->CrushPainSound))
@@ -5042,6 +5054,15 @@ void PIT_FloorDrop (AActor *thing, FChangePosition *cpos)
 			P_CheckFakeFloorTriggers (thing, oldz);
 		}
 	}
+	else if ((thing->z != oldfloorz && !(thing->flags & MF_NOLIFTDROP)))
+	{
+		fixed_t oldz = thing->z;
+		if ((thing->flags & MF_NOGRAVITY) && (thing->flags6 & MF6_RELATIVETOFLOOR))
+		{
+			thing->z = thing->z - oldfloorz + thing->floorz;
+			P_CheckFakeFloorTriggers (thing, oldz);
+		}
+	}
 }
 
 //=============================================================================
@@ -5053,6 +5074,7 @@ void PIT_FloorDrop (AActor *thing, FChangePosition *cpos)
 void PIT_FloorRaise (AActor *thing, FChangePosition *cpos)
 {
 	fixed_t oldfloorz = thing->floorz;
+	fixed_t oldz = thing->z;
 
 	P_AdjustFloorCeil (thing, cpos);
 
@@ -5067,22 +5089,30 @@ void PIT_FloorRaise (AActor *thing, FChangePosition *cpos)
 			return; // do not move bridge things
 		}
 		intersectors.Clear ();
-		fixed_t oldz = thing->z;
 		thing->z = thing->floorz;
-		switch (P_PushUp (thing, cpos))
+	}
+	else
+	{
+		if((thing->flags & MF_NOGRAVITY) && (thing->flags6 & MF6_RELATIVETOFLOOR))
 		{
-		default:
-			P_CheckFakeFloorTriggers (thing, oldz);
-			break;
-		case 1:
-			P_DoCrunch (thing, cpos);
-			P_CheckFakeFloorTriggers (thing, oldz);
-			break;
-		case 2:
-			P_DoCrunch (thing, cpos);
-			thing->z = oldz;
-			break;
+			intersectors.Clear ();
+			thing->z = thing->z - oldfloorz + thing->floorz;
 		}
+		else return;
+	}
+	switch (P_PushUp (thing, cpos))
+	{
+	default:
+		P_CheckFakeFloorTriggers (thing, oldz);
+		break;
+	case 1:
+		P_DoCrunch (thing, cpos);
+		P_CheckFakeFloorTriggers (thing, oldz);
+		break;
+	case 2:
+		P_DoCrunch (thing, cpos);
+		thing->z = oldz;
+		break;
 	}
 }
